+++ apps/web/docs/editor/Preview.ts
@@ -1,34 +1,25 @@
 import { type DomphyElement, type State, ElementNode } from '@domphy/core'
-import { themeApply } from '@domphy/theme'
 import { moduleMap } from './Modules'
 import { Render } from './Render.js'
 import { transformCode } from './transformCode.js'
 
-export function Preview(code: State<string>, isDark: State<boolean>, hasGrid: State<boolean>, error: State<string>): DomphyElement<'div'> {
+export function Preview(code: State<string>, isDark: State<boolean>, hasGrid: State<boolean>, error: State<string>, shadowHost: HTMLElement, previewContainer: HTMLElement): DomphyElement<'div'> {
   return {
     div: [],
     _onMount: (node) => {
       const dom = node.domElement as HTMLElement
-      const shadow = dom.attachShadow({ mode: 'open' })
-      const container = document.createElement('div')
-      container.style.flex = '1'
-      const themeTag = document.createElement('style')
-      themeTag.id = "domphy-themes"
-      shadow.append(themeTag, container)
-      themeApply(themeTag)
+      dom.appendChild(shadowHost)
       let newNode: ElementNode | null = null
 
       const update = (val: string) => {
-        container.textContent = ""
+        previewContainer.textContent = ""
         try {
           if (newNode) newNode.remove()
           const fn = new Function('__modules__', transformCode(val))
           const el = fn(moduleMap)
           if (!el) return
           newNode = new ElementNode(Render(el, isDark, hasGrid))
-
-          newNode.render(container)
-
+          newNode.render(previewContainer)
         } catch (e: any) {
           error.set(e.message)
         }
@@ -40,7 +31,6 @@ export function Preview(code: State<string>, isDark: State<boolean>, hasGrid: St
       flex: 1,
       display: 'flex',
       flexDirection: 'column',
-      overflow:"auto",
     },
   }
 }