+++ packages/core/CHANGELOG.md
@@ -4,3 +4,5 @@
 - Initial release
 ## 0.1.7
 - fix listener type
+## 0.1.8
+- move _notofier from AttributeList to ElementAttribute