+++ packages/theme/CHANGELOG.md
@@ -4,5 +4,7 @@
 - Initial release
 ## 0.1.6
 - add themeColorToken
-## 0.1.9
+## 0.1.10
 - use chromametry palette
+## 0.1.12
+- big change - reactive dataTheme for darkBias